+++ lib/dialog_handler.dart
@@ -4,3 +4,4 @@ export 'package:dialog_handler/src/dialog_handler.dart';
 export 'package:dialog_handler/src/dialog_manager.dart';
 export 'package:dialog_handler/src/models/_export_.dart';
 export 'package:dialog_handler/src/utils/_export_.dart';
+export 'package:dialog_handler/src/widgets_wrappers/liquid_glass.dart';